Add From<RamError> for HypercoreError
